Implementing both the Artificial Intelligence algorithms
Both artificial intelligence algorithms have been implemented. The first algorithm is designed to predict the ball's location on the y-axis and set the paddle's position to that point before the other player scores a point. The second algorithm is designed to follow the ball's location on the y-axis and continuously set the paddle's position to that point. Both of the algorithms are buggy and not working the way they are intended to. The first one is predicting the location correctly but mirroring the paddle position relative to the game screen at random points of time for unknown reasons. The second one is following the location correctly but has not been accounted for the paddle and ball width which causes it to miss the ball at random points of time.