FINALLY!!!!!!
Thanks a lot, MR. KEVIN WOOD. I couldn't have done any of it without you. Have I lately told you that you are a genius!!!!! (  No, but I would surely like to hear it  :))  ). You are the best teacher I ever had and even though we spent most of the time in remote-learning because of Covid-19, we still had a pretty awesome (and unique) bond. The best part of the entire day was that you were even more excited than me to make the AI work and even though we couldn't figure out what the problem was, you still didn't give up at all and finally made the AI work. You never told me to fix the problem all by myself and you were always there when I needed help. I vividly remember both of our expressions when the program worked perfectly. You did it, Mr. Wood. You would forever be an inspiration to me that you should never ever give up on yourself and show perseverance at all times. It is the key to solving all of your problems and when you do solve that problem, that feeling at that very moment is the best thing in the world.
I hope I could spend more time with you and learn from you. I would love and hope that we remain in contact. I am writing this so that whenever future me sees this, he remembers the exact feeling I had at 2:19 PM on January 28th, 2021 and also remembers to be in touch with you.

Thank You, Mr. Wood
Regards
Keshav Mehndiratta
 
Changed the prediction algorithm of the AI to prevent it from mirroring the y-axis of the ball. The AI works perfectly now. Thank you, Mr. Wood. 




    private double setAITarget()
    {
        double x=ball.x;
        double y=ball.y;
        double dx=ball.xVelocity;
        double dy=ball.yVelocity;
        while (x>0)
        {
            if (dy<0)
            {
                x-=Math.abs(y/dy*dx);
                y=0;
                dy=-dy;
            } else {
                x-=Math.abs((GAME_HEIGHT-y)/dy*dx);
                y=GAME_HEIGHT;
                dy=-dy;
            }
        }
        if (dy<0)
        {
            y=GAME_HEIGHT+x/dx*dy;
        } else {
            y=x/dx*dy;
        }
        return y;
    }