Added UI(User Interface)
Made the game user-friendly by adding a UI to allow users to choose players, win score, and read instructions. Predicting AI does not work for the other side at all and can be added in the future. Invincible AI is added as an alternative to control its speed which results in the flickering of the paddle and can be fixed in the future. Everything else looks perfect!!!!!!!